Move readme section, fix composer.json.
Moves a chuck of the readme to the new loadsys/puphpet-release-composer-installer project.

Corrects a typo in the composer.json and adds a temporary working directory to hook into my local copy of the puphpet-release-composer-installer project.

This is a local "WIP" commit since all changes have to be committed to git for composer to be able to use this folder as a `repository` source.